requirements.txt
requests
BeautifulSoup4
pymongo
fastapi
uvicorn
anthropic
google-generativeai
certifi